[front] chore: rename utils webtools to just websearch
